improved test chamber sign
