+++ src/main/java/com/jacksonplayz/thinkingwithaperture/init/ModBlocks.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 import com.google.common.collect.Lists;
 import com.jacksonplayz.thinkingwithaperture.blocks.BlockBase;
-import com.jacksonplayz.thinkingwithaperture.blocks.BlockSignage;
+import com.jacksonplayz.thinkingwithaperture.blocks.BlockTestChamberSign;
 import com.jacksonplayz.thinkingwithaperture.blocks.BlockSurface;
 import com.jacksonplayz.thinkingwithaperture.items.ItemBlockBase;
 
@@ -21,16 +21,14 @@ public class ModBlocks
     public static final Block WHITE_PORTAL_SURFACE;
     public static final Block METAL_SURFACE;
     public static final Block STONE_FLOOR;
-    public static final Block SIGNAGE;
-    public static final Block SIGNAGE_TOP;
+    public static final Block TEST_CHAMBER_SIGN;
 
     static
     {
         WHITE_PORTAL_SURFACE = new BlockSurface(Material.ROCK, true, "white_portal_surface");
         METAL_SURFACE = new BlockSurface(Material.IRON,false, "metal_surface");
         STONE_FLOOR = new BlockSurface(Material.ROCK, true, "stone_floor");
-        SIGNAGE = new BlockSignage("signage", false);
-        SIGNAGE_TOP = new BlockSignage("signage_top", true);
+        TEST_CHAMBER_SIGN = new BlockTestChamberSign("test_chamber_sign");
 
         register();
     }
@@ -40,8 +38,7 @@ public class ModBlocks
         registerFullBlock(WHITE_PORTAL_SURFACE);
         registerFullBlock(METAL_SURFACE);
         registerFullBlock(STONE_FLOOR);
-        registerFullBlock(SIGNAGE);
-        registerFullBlock(SIGNAGE_TOP);
+        registerFullBlock(TEST_CHAMBER_SIGN);
     }
 
     public static void registerFullBlock(Block block)
