+++ src/main/resources/assets/twa/blockstates/test_chamber_sign.json
@@ -1,0 +1,26 @@
+{
+	"forge_marker": 1,
+	"variants": {
+		"facing": {
+			"north": {
+				"y": 180
+			},
+			"east": {
+				"y": 270
+			},
+			"south": {
+			},
+			"west": {
+				"y": 90
+			}
+		},
+		"top": {
+			"true": {
+				"model": "twa:test_chamber_sign_top"
+			},
+			"false": {
+				"model": "twa:test_chamber_sign_bottom"
+			}
+		}
+	}
+}
