--- src/main/resources/assets/twa/models/block/signage_bottom.json
@@ -1,65 +1,0 @@
-{
-	"credit": "Made with Blockbench",
-	"textures": {
-		"particle": "twa:models/signage/signage_bottom_icons",
-		"0": "twa:models/signage/signage_bottom_icons"
-	},
-	"elements": [
-		{
-			"name": "rightBar",
-			"from": [-1, 0, 14.75],
-			"to": [0, 16, 16],
-			"faces": {
-				"north": {"uv": [15.25, 0, 16, 15.25], "texture": "#0"},
-				"east": {"uv": [15.25, 0, 16, 15.25], "texture": "#0"},
-				"west": {"uv": [15.25, 0, 16, 15.25], "texture": "#0"}
-			}
-		},
-		{
-			"name": "screen",
-			"from": [0, 0, 15],
-			"to": [16, 16, 16],
-			"faces": {
-				"north": {"uv": [0.75, 0, 15.25, 15.25], "texture": "#0"}
-			}
-		},
-		{
-			"name": "leftBar",
-			"from": [16, 0, 14.75],
-			"to": [17, 16, 16],
-			"faces": {
-				"north": {"uv": [0, 0, 0.75, 15.25], "texture": "#0"},
-				"east": {"uv": [0, 0, 0.75, 15.25], "texture": "#0"},
-				"west": {"uv": [0, 0, 0.75, 15.25], "texture": "#0"}
-			}
-		},
-		{
-			"name": "topBar",
-			"from": [-1, -1, 14.5],
-			"to": [17, 0, 16],
-			"faces": {
-				"north": {"uv": [0, 15.25, 16, 16], "texture": "#0"},
-				"east": {"uv": [0, 15.25, 16, 16], "texture": "#0"},
-				"west": {"uv": [0, 15.25, 16, 16], "texture": "#0"},
-				"up": {"uv": [0, 15.25, 16, 16], "texture": "#0"},
-				"down": {"uv": [0, 15.25, 16, 16], "texture": "#0"}
-			}
-		}
-	],
-	"groups": [
-		{
-			"name": "signage_bottom",
-			"children": [
-				{
-					"name": "screenTop",
-					"children": [
-						{
-							"name": "top",
-							"children": [0, 1, 2, 3]
-						}
-					]
-				}
-			]
-		}
-	]
-}
