+++ src/main/resources/assets/twa/models/block/test_chamber_sign_bottom.json
@@ -1,0 +1,50 @@
+{
+    "parent": "block/cube",
+    "textures": {
+        "0": "blocks/concrete_silver",
+        "1": "twa:blocks/test_chamber_sign_bottom",
+        "particle": "blocks/concrete_silver"
+    },
+    "elements": [
+        {
+            "from": [ 0, 1, 0 ], 
+            "to": [ 1, 16, 1 ], 
+            "faces": {
+                "north": { "texture": "#0", "uv": [ 15, 0, 16, 15 ], "cullface": "north" },
+                "east": { "texture": "#0", "uv": [ 0, 0, 1, 15 ] },
+                "south": { "texture": "#0", "uv": [ 0, 0, 1, 15 ] },
+                "west": { "texture": "#0", "uv": [ 15, 0, 16, 15 ], "cullface": "west" }
+            }
+        },
+        {
+            "from": [ 15, 1, 0 ], 
+            "to": [ 16, 16, 1 ], 
+            "faces": {
+                "north": { "texture": "#0", "uv": [ 15, 0, 16, 15 ], "cullface": "north" },
+                "east": { "texture": "#0", "uv": [ 15, 0, 16, 15 ], "cullface": "east" },
+                "south": { "texture": "#0", "uv": [ 15, 0, 16, 15 ] },
+                "west": { "texture": "#0", "uv": [ 0, 0, 1, 15 ] }
+            }
+        },
+        {
+            "from": [ 0, 0, 0 ], 
+            "to": [ 16, 1, 1 ], 
+            "faces": {
+                "north": { "texture": "#0", "uv": [ 0, 15, 16, 16 ], "cullface": "north" },
+                "east": { "texture": "#0", "uv": [ 0, 0, 1, 1 ], "cullface": "east" },
+                "south": { "texture": "#0", "uv": [ 0, 15, 16, 16 ] },
+                "west": { "texture": "#0", "uv": [ 15, 0, 16, 1 ], "cullface": "west" },
+                "up": { "texture": "#0", "uv": [ 0, 15, 16, 16 ] },
+                "down": { "texture": "#0", "uv": [ 0, 0, 16, 1 ], "cullface": "down" }
+            }
+        },
+        {
+            "from": [ 1, 1, 0 ], 
+            "to": [ 15, 16, 0.5 ],
+            "faces": {
+                "north": { "texture": "#0", "uv": [ 1, 1, 15, 16 ], "cullface": "north" },
+                "south": { "texture": "#1", "uv": [ 1, 0, 15, 15 ] }
+            }
+        }
+    ]
+}
