+++ src/main/resources/assets/twa/models/item/energized_potato.json
@@ -1,6 +1,6 @@
 {
-    "parent": "item/generated",
+    "parent": "item/generated", 
     "textures": {
-     "layer0": "twa:items/energized_potato"
+        "layer0": "twa:items/energized_potato"
     }
 }
