+++ src/main/resources/assets/twa/models/item/long_fall_boots.json
@@ -1,6 +1,6 @@
 {
-    "parent": "item/generated",
+    "parent": "item/generated", 
     "textures": {
-     "layer0": "twa:items/long_fall_boots"
+        "layer0": "twa:items/long_fall_boots"
     }
 }
