+++ src/main/resources/assets/twa/models/item/metal_surface.json
@@ -1,3 +1,3 @@
 {
-  "parent": "twa:block/metal_surface"
+    "parent": "twa:block/metal_surface"
 }
