+++ src/main/resources/assets/twa/models/item/potatos_parts.json
@@ -1,6 +1,6 @@
 {
-    "parent": "item/generated",
+    "parent": "item/generated", 
     "textures": {
-     "layer0": "twa:items/potatos_parts"
+        "layer0": "twa:items/potatos_parts"
     }
 }
