+++ src/main/resources/assets/twa/models/item/radio.json
@@ -1,77 +1,121 @@
 {
-	"credit": "Made with Blockbench",
-	"textures": {
-		"0": "twa:items/radio"
-	},
-	"elements": [
-		{
-			"from": [4, 0, 6.5],
-			"to": [12, 6, 9.5],
-			"faces": {
-				"north": {"uv": [1.5, 1.5, 5.5, 4.5], "texture": "#0"},
-				"east": {"uv": [0, 1.5, 1.5, 4.5], "texture": "#0"},
-				"south": {"uv": [7, 1.5, 11, 4.5], "texture": "#0"},
-				"west": {"uv": [5.5, 1.5, 7, 4.5], "texture": "#0"},
-				"up": {"uv": [1.5, 0, 5.5, 1.5], "texture": "#0"},
-				"down": {"uv": [5.5, 0, 9.5, 1.5], "texture": "#0"}
-			}
-		},
-		{
-			"from": [5, 6, 7],
-			"to": [11, 7, 9],
-			"faces": {
-				"north": {"uv": [1, 5.5, 4, 6], "texture": "#0"},
-				"east": {"uv": [0, 5.5, 1, 6], "texture": "#0"},
-				"south": {"uv": [5, 5.5, 8, 6], "texture": "#0"},
-				"west": {"uv": [4, 5.5, 5, 6], "texture": "#0"},
-				"up": {"uv": [1, 4.5, 4, 5.5], "texture": "#0"}
-			}
-		},
-		{
-			"from": [9, 7, 8],
-			"to": [10, 13, 9],
-			"faces": {
-				"north": {"uv": [0.5, 6.5, 1, 9.5], "texture": "#0"},
-				"east": {"uv": [0.5, 6.5, 1, 9.5], "texture": "#0"},
-				"south": {"uv": [0.5, 6.5, 1, 9.5], "texture": "#0"},
-				"west": {"uv": [0.5, 6.5, 1, 9.5], "texture": "#0"},
-				"up": {"uv": [0.5, 6.5, 1, 7], "texture": "#0"}
-			}
-		}
-	],
-	"display": {
-		"thirdperson_righthand": {
-			"rotation": [0, 145, 0],
-			"translation": [0, 4, 1.5],
-			"scale": [0.5, 0.5, 0.5]
-		},
-		"thirdperson_lefthand": {
-			"rotation": [0, 145, 0],
-			"translation": [0, 4, 1.5],
-			"scale": [0.5, 0.5, 0.5]
-		},
-		"firstperson_righthand": {
-			"rotation": [0, 118, 0],
-			"translation": [1.5, 3.25, 1.5],
-			"scale": [0.5, 0.5, 0.5]
-		},
-		"firstperson_lefthand": {
-			"rotation": [0, 118, 0],
-			"translation": [1.5, 3.25, 1.5],
-			"scale": [0.5, 0.5, 0.5]
-		},
-		"ground": {
-			"scale": [0.5, 0.5, 0.5]
-		},
-		"gui": {
-			"rotation": [35, -145, 0],
-			"translation": [0, 1.5, 0]
-		},
-		"head": {
-			"translation": [0, 14.25, 0]
-		},
-		"fixed": {
-			"translation": [0, 0, -1]
-		}
-	}
+    "credit": "Made with Blockbench", 
+    "textures": {
+        "0": "twa:items/radio"
+    }, 
+    "elements": [ {
+            "from": [ 4, 0, 6.5 ], 
+            "to": [ 12, 6, 9.5 ], 
+            "faces": {
+                "north": {
+                    "uv": [ 1.5, 1.5, 5.5, 4.5 ], 
+                    "texture": "#0"
+                }, 
+                "east": {
+                    "uv": [ 0, 1.5, 1.5, 4.5 ], 
+                    "texture": "#0"
+                }, 
+                "south": {
+                    "uv": [ 7, 1.5, 11, 4.5 ], 
+                    "texture": "#0"
+                }, 
+                "west": {
+                    "uv": [ 5.5, 1.5, 7, 4.5 ], 
+                    "texture": "#0"
+                }, 
+                "up": {
+                    "uv": [ 1.5, 0, 5.5, 1.5 ], 
+                    "texture": "#0"
+                }, 
+                "down": {
+                    "uv": [ 5.5, 0, 9.5, 1.5 ], 
+                    "texture": "#0"
+                }
+            }
+        }, {
+            "from": [ 5, 6, 7 ], 
+            "to": [ 11, 7, 9 ], 
+            "faces": {
+                "north": {
+                    "uv": [ 1, 5.5, 4, 6 ], 
+                    "texture": "#0"
+                }, 
+                "east": {
+                    "uv": [ 0, 5.5, 1, 6 ], 
+                    "texture": "#0"
+                }, 
+                "south": {
+                    "uv": [ 5, 5.5, 8, 6 ], 
+                    "texture": "#0"
+                }, 
+                "west": {
+                    "uv": [ 4, 5.5, 5, 6 ], 
+                    "texture": "#0"
+                }, 
+                "up": {
+                    "uv": [ 1, 4.5, 4, 5.5 ], 
+                    "texture": "#0"
+                }
+            }
+        }, {
+            "from": [ 9, 7, 8 ], 
+            "to": [ 10, 13, 9 ], 
+            "faces": {
+                "north": {
+                    "uv": [ 0.5, 6.5, 1, 9.5 ], 
+                    "texture": "#0"
+                }, 
+                "east": {
+                    "uv": [ 0.5, 6.5, 1, 9.5 ], 
+                    "texture": "#0"
+                }, 
+                "south": {
+                    "uv": [ 0.5, 6.5, 1, 9.5 ], 
+                    "texture": "#0"
+                }, 
+                "west": {
+                    "uv": [ 0.5, 6.5, 1, 9.5 ], 
+                    "texture": "#0"
+                }, 
+                "up": {
+                    "uv": [ 0.5, 6.5, 1, 7 ], 
+                    "texture": "#0"
+                }
+            }
+        } ], 
+    "display": {
+        "thirdperson_righthand": {
+            "rotation": [ 0, 145, 0 ], 
+            "translation": [ 0, 4, 1.5 ], 
+            "scale": [ 0.5, 0.5, 0.5 ]
+        }, 
+        "thirdperson_lefthand": {
+            "rotation": [ 0, 145, 0 ], 
+            "translation": [ 0, 4, 1.5 ], 
+            "scale": [ 0.5, 0.5, 0.5 ]
+        }, 
+        "firstperson_righthand": {
+            "rotation": [ 0, 118, 0 ], 
+            "translation": [ 1.5, 3.25, 1.5 ], 
+            "scale": [ 0.5, 0.5, 0.5 ]
+        }, 
+        "firstperson_lefthand": {
+            "rotation": [ 0, 118, 0 ], 
+            "translation": [ 1.5, 3.25, 1.5 ], 
+            "scale": [ 0.5, 0.5, 0.5 ]
+        }, 
+        "ground": {
+            "scale": [ 0.5, 0.5, 0.5 ]
+        }, 
+        "gui": {
+            "rotation": [ 35, -145, 0 ], 
+            "translation": [ 0, 1.5, 0 ]
+        }, 
+        "head": {
+            "translation": [ 0, 14.25, 0 ]
+        }, 
+        "fixed": {
+            "translation": [ 0, 0, -1 ]
+        }
+    }
 }
