+++ src/main/resources/assets/twa/models/item/record_cara_mia_addio.json
@@ -1,5 +1,5 @@
 {
-    "parent": "item/generated",
+    "parent": "item/generated", 
     "textures": {
         "layer0": "twa:items/record_cara_mia_addio"
     }
