+++ src/main/resources/assets/twa/models/item/record_want_you_gone.json
@@ -1,5 +1,5 @@
 {
-    "parent": "item/generated",
+    "parent": "item/generated", 
     "textures": {
         "layer0": "twa:items/record_want_you_gone"
     }
