--- src/main/resources/assets/twa/models/item/signage.json
@@ -1,150 +1,0 @@
-{
-	"credit": "Made with Blockbench",
-	"textures": {
-		"0": "twa:models/signage/signage_bottom_icons",
-		"2": "twa:models/signage/signage_top",
-		"3": "blocks/anvil_base"
-	},
-	"elements": [
-		{
-			"name": "bottomRightBar",
-			"from": [-1, -10, -0.25],
-			"to": [0, 6, 1],
-			"faces": {
-				"north": {"uv": [15.25, 0, 16, 15.25], "texture": "#0"},
-				"east": {"uv": [15.25, 0, 16, 15.25], "texture": "#0"},
-				"south": {"uv": [15.25, 0, 16, 15.25], "texture": "#0"},
-				"west": {"uv": [15.25, 0, 16, 15.25], "texture": "#0"},
-				"up": {"uv": [15.25, 0, 16, 15.25], "texture": "#0"},
-				"down": {"uv": [15.25, 0, 16, 15.25], "texture": "#0"}
-			}
-		},
-		{
-			"name": "screenBottom",
-			"from": [0, -10, 0],
-			"to": [16, 6, 1],
-			"faces": {
-				"north": {"uv": [0.75, 0, 15.25, 15.25], "texture": "#0"},
-				"south": {"uv": [0, 0, 16, 16], "texture": "#3"}
-			}
-		},
-		{
-			"name": "bottomLeftBar",
-			"from": [16, -10, -0.25],
-			"to": [17, 6, 1],
-			"faces": {
-				"north": {"uv": [0, 0, 0.75, 15.25], "texture": "#0"},
-				"east": {"uv": [0, 0, 0.75, 15.25], "texture": "#0"},
-				"south": {"uv": [0, 0, 0.75, 15.25], "texture": "#0"},
-				"west": {"uv": [0, 0, 0.75, 15.25], "texture": "#0"},
-				"up": {"uv": [0, 0, 0.75, 15.25], "texture": "#0"},
-				"down": {"uv": [0, 0, 0.75, 15.25], "texture": "#0"}
-			}
-		},
-		{
-			"name": "bottomBar",
-			"from": [-1, -11, -0.5],
-			"to": [17, -10, 1],
-			"faces": {
-				"north": {"uv": [0, 15.25, 16, 16], "texture": "#0"},
-				"east": {"uv": [0, 15.25, 16, 16], "texture": "#0"},
-				"south": {"uv": [0, 15.25, 16, 16], "texture": "#0"},
-				"west": {"uv": [0, 15.25, 16, 16], "texture": "#0"},
-				"up": {"uv": [0, 15.25, 16, 16], "texture": "#0"},
-				"down": {"uv": [0, 15.25, 16, 16], "texture": "#0"}
-			}
-		},
-		{
-			"name": "topRightBar",
-			"from": [-1, 6, -0.25],
-			"to": [0, 22, 1],
-			"faces": {
-				"north": {"uv": [15.25, 0.75, 16, 16], "texture": "#2"},
-				"east": {"uv": [15.25, 0.75, 16, 16], "texture": "#2"},
-				"south": {"uv": [15.25, 0.75, 16, 16], "texture": "#2"},
-				"west": {"uv": [15.25, 0.75, 16, 16], "texture": "#2"},
-				"up": {"uv": [15.25, 0.75, 16, 16], "texture": "#2"},
-				"down": {"uv": [15.25, 0.75, 16, 16], "texture": "#2"}
-			}
-		},
-		{
-			"name": "screenTop",
-			"from": [0, 6, 0],
-			"to": [16, 22, 1],
-			"faces": {
-				"north": {"uv": [0.75, 0.75, 15.25, 16], "texture": "#2"},
-				"south": {"uv": [0, 0, 16, 16], "texture": "#3"}
-			}
-		},
-		{
-			"name": "topLeftBar",
-			"from": [16, 6, -0.25],
-			"to": [17, 22, 1],
-			"faces": {
-				"north": {"uv": [0, 0.75, 0.75, 16], "texture": "#2"},
-				"east": {"uv": [0, 0.75, 0.75, 16], "texture": "#2"},
-				"south": {"uv": [0, 0.75, 0.75, 16], "texture": "#2"},
-				"west": {"uv": [0, 0.75, 0.75, 16], "texture": "#2"},
-				"up": {"uv": [0, 0.75, 0.75, 16], "texture": "#2"},
-				"down": {"uv": [0, 0.75, 0.75, 16], "texture": "#2"}
-			}
-		},
-		{
-			"name": "topBar",
-			"from": [-1, 22, -0.5],
-			"to": [17, 23, 1],
-			"faces": {
-				"north": {"uv": [0, 0, 16, 0.75], "texture": "#2"},
-				"east": {"uv": [0, 0, 16, 0.75], "texture": "#2"},
-				"south": {"uv": [0, 0, 16, 0.75], "texture": "#2"},
-				"west": {"uv": [0, 0, 16, 0.75], "texture": "#2"},
-				"up": {"uv": [0, 0, 16, 0.75], "texture": "#2"},
-				"down": {"uv": [0, 0, 16, 0.75], "texture": "#2"}
-			}
-		}
-	],
-	"display": {
-		"thirdperson_righthand": {
-			"rotation": [0, 180, 0],
-			"translation": [0, 4.25, -1],
-			"scale": [0.25, 0.25, 0.25]
-		},
-		"thirdperson_lefthand": {
-			"rotation": [0, 180, 0],
-			"translation": [0, 4.25, -1],
-			"scale": [0.25, 0.25, 0.25]
-		},
-		"firstperson_righthand": {
-			"rotation": [-1, 129, 0],
-			"translation": [5.25, 2.25, -2.75],
-			"scale": [0.25, 0.25, 0.5]
-		},
-		"firstperson_lefthand": {
-			"rotation": [-1, 129, 0],
-			"translation": [5.25, 2.25, -2.75],
-			"scale": [0.25, 0.25, 0.5]
-		},
-		"ground": {
-			"translation": [0, 1.25, 0],
-			"scale": [0.25, 0.25, 0.25]
-		},
-		"gui": {
-			"rotation": [35, -150, 0],
-			"translation": [-1.35, 2.5, 0],
-			"scale": [0.425, 0.425, 0.425]
-		},
-		"head": {
-			"translation": [0, -5.5, 2]
-		},
-		"fixed": {
-			"translation": [0, 1, 3.5],
-			"scale": [0.5, 0.5, 0.5]
-		}
-	},
-	"groups": [
-		{
-			"name": "screen",
-			"children": [0, 1, 2, 3, 4, 5, 6, 7]
-		}
-	]
-}
