+++ src/main/resources/assets/twa/models/item/test_chamber_sign.json
@@ -1,0 +1,222 @@
+{
+    "textures": {
+        "border": "blocks/concrete_silver", 
+        "bottomScreen": "twa:blocks/test_chamber_sign_bottom", 
+        "topScreen": "twa:blocks/test_chamber_sign_top", 
+        "particle": "blocks/concrete_silver"
+    }, 
+    "display": {
+        "gui": {
+            "rotation": [ 30, 45, 0 ], 
+            "translation": [ 2.5, -4, 0 ], 
+            "scale": [ 0.425, 0.425, 0.425 ]
+        }, 
+        "ground": {
+            "rotation": [ 0, 180, 0 ], 
+            "translation": [ 0, 3, -2 ], 
+            "scale": [ 0.25, 0.25, 0.25 ]
+        }, 
+        "fixed": {
+            "rotation": [ 0, 180, 0 ], 
+            "translation": [ 0, 0, 0 ], 
+            "scale": [ 0.5, 0.5, 0.5 ]
+        }, 
+        "thirdperson_righthand": {
+            "rotation": [ 0, 0, 0 ], 
+            "translation": [ 0, 1.5, 2.5 ], 
+            "scale": [ 0.25, 0.25, 0.25 ]
+        }, 
+        "firstperson_righthand": {
+            "rotation": [ 0, 315, 0 ], 
+            "translation": [ 0, 0, 0 ], 
+            "scale": [ 0.40, 0.40, 0.40 ]
+        }
+    }, 
+    "elements": [ {
+            "from": [ 0, 0, 0 ], 
+            "to": [ 16, 1, 1 ], 
+            "faces": {
+                "north": {
+                    "texture": "#border", 
+                    "uv": [ 0, 15, 16, 16 ], 
+                    "cullface": "north"
+                }, 
+                "east": {
+                    "texture": "#border", 
+                    "uv": [ 0, 0, 1, 1 ], 
+                    "cullface": "east"
+                }, 
+                "south": {
+                    "texture": "#border", 
+                    "uv": [ 0, 15, 16, 16 ]
+                }, 
+                "west": {
+                    "texture": "#border", 
+                    "uv": [ 15, 0, 16, 1 ], 
+                    "cullface": "west"
+                }, 
+                "up": {
+                    "texture": "#border", 
+                    "uv": [ 0, 15, 16, 16 ]
+                }, 
+                "down": {
+                    "texture": "#border", 
+                    "uv": [ 0, 0, 16, 1 ], 
+                    "cullface": "down"
+                }
+            }
+        }, {
+            "from": [ 0, 1, 0 ], 
+            "to": [ 1, 16, 1 ], 
+            "faces": {
+                "north": {
+                    "texture": "#border", 
+                    "uv": [ 15, 0, 16, 15 ], 
+                    "cullface": "north"
+                }, 
+                "east": {
+                    "texture": "#border", 
+                    "uv": [ 0, 0, 1, 15 ]
+                }, 
+                "south": {
+                    "texture": "#border", 
+                    "uv": [ 0, 0, 1, 15 ]
+                }, 
+                "west": {
+                    "texture": "#border", 
+                    "uv": [ 15, 0, 16, 15 ], 
+                    "cullface": "west"
+                }
+            }
+        }, {
+            "from": [ 15, 1, 0 ], 
+            "to": [ 16, 16, 1 ], 
+            "faces": {
+                "north": {
+                    "texture": "#border", 
+                    "uv": [ 15, 0, 16, 15 ], 
+                    "cullface": "north"
+                }, 
+                "east": {
+                    "texture": "#border", 
+                    "uv": [ 15, 0, 16, 15 ], 
+                    "cullface": "east"
+                }, 
+                "south": {
+                    "texture": "#border", 
+                    "uv": [ 15, 0, 16, 15 ]
+                }, 
+                "west": {
+                    "texture": "#border", 
+                    "uv": [ 0, 0, 1, 15 ]
+                }
+            }
+        }, {
+            "from": [ 1, 1, 0 ], 
+            "to": [ 15, 16, 0.5 ], 
+            "faces": {
+                "north": {
+                    "texture": "#border", 
+                    "uv": [ 1, 0, 15, 15 ], 
+                    "cullface": "north"
+                }, 
+                "south": {
+                    "texture": "#bottomScreen", 
+                    "uv": [ 1, 0, 15, 15 ]
+                }
+            }
+        }, {
+            "from": [ 0, 16, 0 ], 
+            "to": [ 1, 31, 1 ], 
+            "faces": {
+                "north": {
+                    "texture": "#border", 
+                    "uv": [ 15, 1, 16, 16 ], 
+                    "cullface": "north"
+                }, 
+                "east": {
+                    "texture": "#border", 
+                    "uv": [ 0, 1, 1, 16 ]
+                }, 
+                "south": {
+                    "texture": "#border", 
+                    "uv": [ 0, 1, 1, 16 ]
+                }, 
+                "west": {
+                    "texture": "#border", 
+                    "uv": [ 15, 1, 16, 16 ], 
+                    "cullface": "west"
+                }
+            }
+        }, {
+            "from": [ 15, 16, 0 ], 
+            "to": [ 16, 31, 1 ], 
+            "faces": {
+                "north": {
+                    "texture": "#border", 
+                    "uv": [ 15, 1, 16, 16 ], 
+                    "cullface": "north"
+                }, 
+                "east": {
+                    "texture": "#border", 
+                    "uv": [ 15, 1, 16, 16 ], 
+                    "cullface": "east"
+                }, 
+                "south": {
+                    "texture": "#border", 
+                    "uv": [ 15, 1, 16, 16 ]
+                }, 
+                "west": {
+                    "texture": "#border", 
+                    "uv": [ 0, 1, 1, 16 ]
+                }
+            }
+        }, {
+            "from": [ 0, 31, 0 ], 
+            "to": [ 16, 32, 1 ], 
+            "faces": {
+                "north": {
+                    "texture": "#border", 
+                    "uv": [ 0, 15, 16, 16 ], 
+                    "cullface": "north"
+                }, 
+                "east": {
+                    "texture": "#border", 
+                    "uv": [ 0, 0, 1, 1 ], 
+                    "cullface": "east"
+                }, 
+                "south": {
+                    "texture": "#border", 
+                    "uv": [ 0, 15, 16, 16 ]
+                }, 
+                "west": {
+                    "texture": "#border", 
+                    "uv": [ 15, 0, 16, 1 ], 
+                    "cullface": "west"
+                }, 
+                "up": {
+                    "texture": "#border", 
+                    "uv": [ 0, 15, 16, 16 ]
+                }, 
+                "down": {
+                    "texture": "#border", 
+                    "uv": [ 0, 0, 16, 1 ], 
+                    "cullface": "down"
+                }
+            }
+        }, {
+            "from": [ 1, 16, 0 ], 
+            "to": [ 15, 31, 0.5 ], 
+            "faces": {
+                "north": {
+                    "texture": "#border", 
+                    "uv": [ 1, 1, 15, 16 ], 
+                    "cullface": "north"
+                }, 
+                "south": {
+                    "texture": "#topScreen", 
+                    "uv": [ 1, 1, 15, 16 ]
+                }
+            }
+        } ]
+}
