+++ src/main/resources/assets/twa/models/item/turret_fat.json
@@ -1,6 +1,6 @@
 {
-    "parent": "item/generated",
+    "parent": "item/generated", 
     "textures": {
-     "layer0": "twa:items/big_turret"
+        "layer0": "twa:items/big_turret"
     }
 }
