+++ src/main/resources/assets/twa/models/item/turret_normal.json
@@ -1,6 +1,6 @@
 {
-    "parent": "item/generated",
+    "parent": "item/generated", 
     "textures": {
-     "layer0": "twa:items/turret"
+        "layer0": "twa:items/turret"
     }
 }
