+++ src/main/resources/assets/twa/models/item/weighted_companion_cube.json
@@ -1,33 +1,66 @@
 {
-	"credit": "Made by MCvinnyq",
-	"parent": "block/block",
-	"textures": {
-		"0": "twa:entity/cube/companion_cube"
-	},
-	"elements": [
-		{
-			"from": [0, 0, 0],
-			"to": [16, 16, 16],
-			"faces": {
-				"north": {"uv": [4, 4, 8, 8], "texture": "#0"},
-				"east": {"uv": [0, 4, 4, 8], "texture": "#0"},
-				"south": {"uv": [12, 4, 16, 8], "texture": "#0"},
-				"west": {"uv": [8, 4, 12, 8], "texture": "#0"},
-				"up": {"uv": [4, 0, 8, 4], "texture": "#0"},
-				"down": {"uv": [8, 0, 12, 4], "texture": "#0"}
-			}
-		},
-		{
-			"from": [0.25, 0.25, 0.25],
-			"to": [15.75, 15.75, 15.79],
-			"faces": {
-				"north": {"uv": [3.75, 8, 7.5, 11.75], "texture": "#0"},
-				"east": {"uv": [3.75, 8, 7.5, 11.75], "texture": "#0"},
-				"south": {"uv": [3.75, 8, 7.5, 11.75], "texture": "#0"},
-				"west": {"uv": [3.75, 8, 7.5, 11.75], "texture": "#0"},
-				"up": {"uv": [3.75, 8, 7.5, 11.75], "texture": "#0"},
-				"down": {"uv": [3.75, 8, 7.5, 11.75], "texture": "#0"}
-			}
-		}
-	]
+    "credit": "Made by MCvinnyq", 
+    "parent": "block/block", 
+    "textures": {
+        "0": "twa:entity/cube/companion_cube"
+    }, 
+    "elements": [ {
+            "from": [ 0, 0, 0 ], 
+            "to": [ 16, 16, 16 ], 
+            "faces": {
+                "north": {
+                    "uv": [ 4, 4, 8, 8 ], 
+                    "texture": "#0"
+                }, 
+                "east": {
+                    "uv": [ 0, 4, 4, 8 ], 
+                    "texture": "#0"
+                }, 
+                "south": {
+                    "uv": [ 12, 4, 16, 8 ], 
+                    "texture": "#0"
+                }, 
+                "west": {
+                    "uv": [ 8, 4, 12, 8 ], 
+                    "texture": "#0"
+                }, 
+                "up": {
+                    "uv": [ 4, 0, 8, 4 ], 
+                    "texture": "#0"
+                }, 
+                "down": {
+                    "uv": [ 8, 0, 12, 4 ], 
+                    "texture": "#0"
+                }
+            }
+        }, {
+            "from": [ 0.25, 0.25, 0.25 ], 
+            "to": [ 15.75, 15.75, 15.79 ], 
+            "faces": {
+                "north": {
+                    "uv": [ 3.75, 8, 7.5, 11.75 ], 
+                    "texture": "#0"
+                }, 
+                "east": {
+                    "uv": [ 3.75, 8, 7.5, 11.75 ], 
+                    "texture": "#0"
+                }, 
+                "south": {
+                    "uv": [ 3.75, 8, 7.5, 11.75 ], 
+                    "texture": "#0"
+                }, 
+                "west": {
+                    "uv": [ 3.75, 8, 7.5, 11.75 ], 
+                    "texture": "#0"
+                }, 
+                "up": {
+                    "uv": [ 3.75, 8, 7.5, 11.75 ], 
+                    "texture": "#0"
+                }, 
+                "down": {
+                    "uv": [ 3.75, 8, 7.5, 11.75 ], 
+                    "texture": "#0"
+                }
+            }
+        } ]
 }
