+++ src/main/resources/assets/twa/models/item/white_portal_surface.json
@@ -1,3 +1,3 @@
 {
-  "parent": "twa:block/white_portal_surface"
+    "parent": "twa:block/white_portal_surface"
 }
