+++ src/main/resources/assets/twa/textures/blocks/test_chamber_sign_bottom_ref.png
